refactor(engine): extract numeric kernel from engine.py (#843)
Split src/tinycta/engine.py into a thin Polars-facing orchestration layer
and a new pure-NumPy numeric kernel (src/tinycta/_kernel.py) holding
_risk_position, _update_profit_variance, _denominator_is_degenerate, and the
forward_walk loop (previously the Engine._forward_walk method, now a free
function taking `shrink` as a parameter).

engine.py drops from 277 -> 160 LOC and its maintainability index rises from
A(60.8) to A(69.1); _kernel.py is A(72.7). Coverage stays 100% (line+branch).
Kernel-function imports in the two engine test modules now point at
tinycta._kernel.

Closes #843
